Program activity tests now working
