+++ dataactvalidator/config/awardFinancialRules.csv
@@ -24,4 +24,4 @@ MIN_LENGTH,TRUE,ObjectClass,4,,"Checks if object class is 4 digits, if so will c
 CHECK_PREFIX,TRUE,ObjectClass,ByDirectReimbursableFundingSource,"{""1"":""D"",""2"":""R""}",Check that 1 digit prefix is consistent with value for ByDirectReimbursableFundingSource,B11_C3,2
 RULE_IF,TRUE,ObjectClass,B11_C3,B11_C2,"If object class is 4 digits, check that 1 digit prefix is consistent with value for ByDirectReimbursableFundingSource",B11_C4,1
 SUM_FIELDS,FALSE,,DeobligationsRecoveriesRefundsOfPriorYearByAward_CPE,"ussgl487100_downwardadjustmentsofprioryearunpaidundeliveredordersobligationsrecoveries_cpe, ussgl487200_downwardadjustmentsofprioryearprepaidadvancedundeliveredordersobligationsrefundscollected_cpe, ussgl497100_downwardadjustmentsofprioryearunpaiddeliveredordersobligationsrecoveries_cpe, ussgl497200_downwardadjustmentsofprioryearpaiddeliveredordersobligationsrefundscollected_cpe",DeobligationsRecoveriesRefundsOfPriorYearByAward_CPE = USSGL(4871+ 4872 + 4971 + 4972). This applies to the award level.,C18,4
-SET_EXISTS_IN_TABLE,FALSE,,ProgramActivity,"{""beginningperiodofavailability"":""budget_year"", ""agencyidentifier"":""agency_id"", ""allocationtransferagencyidentifier"":""allocation_transfer_id"",""mainaccountcode"":""account_number"",""programactivitycode"":""program_activity_code"",""programactivityname"":""program_activity_name""}",Must be a valid program activity name and code for the corresponding TAS/TAFS as defined in Section 82 of OMB Circular A-11.,B9,4
+SET_EXISTS_IN_TABLE,FALSE,,ProgramActivity,"{""beginningperiodofavailability"":{""target_field"":""budget_year"",""skip_if_below"":2016}, ""agencyidentifier"":{""target_field"":""agency_id"",""pad_to_length"":3}, ""allocationtransferagencyidentifier"":{""target_field"":""allocation_transfer_id"",""pad_to_length"":3},""mainaccountcode"":{""target_field"":""account_number"",""pad_to_length"":4},""programactivitycode"":{""target_field"":""program_activity_code"",""pad_to_length"":4},""programactivityname"":{""target_field"":""program_activity_name""}}",Must be a valid program activity name and code for the corresponding TAS/TAFS as defined in Section 82 of OMB Circular A-11.,B9,4
