+++ dataactvalidator/filestreaming/loaderUtils.py
@@ -1,5 +1,6 @@
 import csv
 from dataactvalidator.filestreaming.fieldCleaner import FieldCleaner
+from dataactvalidator.validation_handlers.validator import Validator
 
 class LoaderUtils:
     @staticmethod
@@ -72,7 +73,7 @@ class LoaderUtils:
                     options = fieldOptions[field]
                     if "pad_to_length" in options:
                         padLength = options["pad_to_length"]
-                        row[field] = cls.padToLength(row[field],padLength)
+                        row[field] = Validator.padToLength(row[field],padLength)
                     if "skip_duplicates" in options:
                         if len(row[field].strip()) == 0 or row[field] in valuePresent[field]:
                             # Value not provided or already exists, skip it
@@ -85,26 +86,4 @@ class LoaderUtils:
                     interface.session.merge(record)
             interface.session.commit()
 
-    @staticmethod
-    def padToLength(data,padLength):
-        """ Pad data with leading zeros
 
-        Args:
-            data: string to be padded
-            padLength: length of string after padding
-
-        Returns:
-            padded string of length padLength
-        """
-        data = data.strip()
-
-        if len(data) < padLength:
-            numZeros = padLength - len(data)
-            zeros = "0" * numZeros
-            result = zeros + str(data)
-            return result
-        elif len(data) > padLength:
-            raise ValueError("".join(["Value is too long: ",str(data)]))
-        else:
-            # Data is correct length already
-            return data
