+++ tests/awardFinancialMixed.csv
@@ -1,13 +1,13 @@
 allocationtransferagencyidentifier,agencyidentifier,beginningperiodofavailability,endingperiodofavailability,availabilitytypecode,mainaccountcode,subaccountcode,piid,parentawardid,fain,uri,programactivityname,programactivitycode,objectclass,bydirectreimbursablefundingsource,obligationsincurredtotalbyaward_cpe,obligationsundeliveredordersunpaidtotal_fyb,obligationsundeliveredordersunpaidtotal_cpe,ussgl480100_undeliveredordersobligationsunpaid_fyb,ussgl480100_undeliveredordersobligationsunpaid_cpe,ussgl483100_undeliveredordersobligationstransferredunpaid_cpe,ussgl488100_upwardadjustmentsofprioryearundeliveredordersobligationsunpaid_cpe,obligationsdeliveredordersunpaidtotal_fyb,obligationsdeliveredordersunpaidtotal_cpe,ussgl490100_deliveredordersobligationsunpaid_fyb,ussgl490100_deliveredordersobligationsunpaid_cpe,ussgl493100_deliveredordersobligationstransferredunpaid_cpe,ussgl498100_upwardadjustmentsofprioryeardeliveredordersobligationsunpaid_cpe,grossoutlayamountbyaward_fyb,grossoutlayamountbyaward_cpe,grossoutlaysundeliveredordersprepaidtotal_fyb,grossoutlaysundeliveredordersprepaidtotal_cpe,ussgl480200_undeliveredordersobligationsprepaidadvanced_fyb,ussgl480200_undeliveredordersobligationsprepaidadvanced_cpe,ussgl483200_undeliveredordersobligationstransferredprepaidadvanced_cpe,ussgl488200_upwardadjustmentsofprioryearundeliveredordersobligationsprepaidadvanced_cpe,grossoutlaysdeliveredorderspaidtotal_fyb,grossoutlaysdeliveredorderspaidtotal_cpe,ussgl490200_deliveredordersobligationspaid_cpe,ussgl490800_authorityoutlayednotyetdisbursed_fyb,ussgl490800_authorityoutlayednotyetdisbursed_cpe,ussgl498200_upwardadjustmentsofprioryeardeliveredordersobligationspaid_cpe,transactionobligatedamount,ussgl487100_downwardadjustmentsofprioryearunpaidundeliveredordersobligationsrecoveries_cpe,ussgl497100_downwardadjustmentsofprioryearunpaiddeliveredordersobligationsrecoveries_cpe,ussgl487200_downwardadjustmentsofprioryearprepaidadvancedundeliveredordersobligationsrefundscollected_cpe,ussgl497200_downwardadjustmentsofprioryearpaiddeliveredordersobligationsrefundscollected_cpe,deobligationsrecoveriesrefundsofprioryearbyaward_cpe
-19,72,0,0,X,306,0,ABCD,ABCD,123456798,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,ABCD,ABCD,123456798,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
 ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-19,72,0,0,X,306,0,ABCD,ABCD,123456800,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,ABCD,ABCD,D,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD
-1900,7200,10000,10000,XFA,10950,9999,ABCD,ABCD,123456801,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,23456,25412,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,1,ABCD,ABCD,123456802,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,,6501,6502,6503,6504,6505,6506,6507,6508,6509,6510,6511,6512,6513,6514,6515,6516,6517,6518,6519,6520,6521,6522,6523,6524,6525,6526,6527,6528,6529,6530,6531,6532,6533
-19,72,0,0,X,306,0,ABCD,ABCD,123456803,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,,6500,6500,6500,6500,26000
-19,72,0,0,B,306,0,ABCD,ABCD,123456804,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,R,,,,,,,,,,,,,,,,,,,,,,,,,,,,6500,,,,,
-19,72,0,0,X,306,0,ABCD,ABCD,123456805,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,0,,ABCD,123456806,,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,0,ABCD,ABCD,,,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,0,,ABCD,,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,0,,ABCD,,,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,ABCD,ABCD,123456800,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,ABCD,ABCD,D,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD,ABCD
+1900,7200,10000,10000,XFA,10950,9999,ABCD,ABCD,123456801,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,23456,25412,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,1,ABCD,ABCD,123456802,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,,6501,6502,6503,6504,6505,6506,6507,6508,6509,6510,6511,6512,6513,6514,6515,6516,6517,6518,6519,6520,6521,6522,6523,6524,6525,6526,6527,6528,6529,6530,6531,6532,6533
+0,19,2016,2016,,113,0,ABCD,ABCD,123456803,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,,6500,6500,6500,6500,26000
+0,19,2013,2016,B,999,0,ABCD,ABCD,123456804,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,R,,,,,,,,,,,,,,,,,,,,,,,,,,,,6500,,,,,
+0,19,2016,2016,,113,0,ABCD,ABCD,123456805,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,,ABCD,123456806,,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,ABCD,ABCD,,,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,,ABCD,,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,,ABCD,,,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
