+++ tests/cross_file_C.csv
@@ -1,3 +1,3 @@
 allocationtransferagencyidentifier,agencyidentifier,beginningperiodofavailability,endingperiodofavailability,availabilitytypecode,mainaccountcode,subaccountcode,piid,parentawardid,fain,uri,programactivityname,programactivitycode,objectclass,bydirectreimbursablefundingsource,obligationsincurredtotalbyaward_cpe,obligationsundeliveredordersunpaidtotal_fyb,obligationsundeliveredordersunpaidtotal_cpe,ussgl480100_undeliveredordersobligationsunpaid_fyb,ussgl480100_undeliveredordersobligationsunpaid_cpe,ussgl483100_undeliveredordersobligationstransferredunpaid_cpe,ussgl488100_upwardadjustmentsofprioryearundeliveredordersobligationsunpaid_cpe,obligationsdeliveredordersunpaidtotal_fyb,obligationsdeliveredordersunpaidtotal_cpe,ussgl490100_deliveredordersobligationsunpaid_fyb,ussgl490100_deliveredordersobligationsunpaid_cpe,ussgl493100_deliveredordersobligationstransferredunpaid_cpe,ussgl498100_upwardadjustmentsofprioryeardeliveredordersobligationsunpaid_cpe,grossoutlayamountbyaward_fyb,grossoutlayamountbyaward_cpe,grossoutlaysundeliveredordersprepaidtotal_fyb,grossoutlaysundeliveredordersprepaidtotal_cpe,ussgl480200_undeliveredordersobligationsprepaidadvanced_fyb,ussgl480200_undeliveredordersobligationsprepaidadvanced_cpe,ussgl483200_undeliveredordersobligationstransferredprepaidadvanced_cpe,ussgl488200_upwardadjustmentsofprioryearundeliveredordersobligationsprepaidadvanced_cpe,grossoutlaysdeliveredorderspaidtotal_fyb,grossoutlaysdeliveredorderspaidtotal_cpe,ussgl490200_deliveredordersobligationspaid_cpe,ussgl490800_authorityoutlayednotyetdisbursed_fyb,ussgl490800_authorityoutlayednotyetdisbursed_cpe,ussgl498200_upwardadjustmentsofprioryeardeliveredordersobligationspaid_cpe,transactionobligatedamount,ussgl487100_downwardadjustmentsofprioryearunpaidundeliveredordersobligationsrecoveries_cpe,ussgl497100_downwardadjustmentsofprioryearunpaiddeliveredordersobligationsrecoveries_cpe,ussgl487200_downwardadjustmentsofprioryearprepaidadvancedundeliveredordersobligationsrefundscollected_cpe,ussgl497200_downwardadjustmentsofprioryearpaiddeliveredordersobligationsrefundscollected_cpe,deobligationsrecoveriesrefundsofprioryearbyaward_cpe
-19,72,0,0,X,306,0,ABCD,ABCD,123456789,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
-19,72,0,0,X,306,0,ABCD,ABCD,123456799,12-34-56-78-90-AS-DF-AB-XZ-YW,School improvement grants,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,ABCD,ABCD,123456789,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
+0,19,2016,2016,,113,0,ABCD,ABCD,123456799,12-34-56-78-90-AS-DF-AB-XZ-YW,Overseas Programs,2,1255,D,6500,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,13000,52000,6500,26000,6500,6500,6500,6500,6500,26000,6500,6500,6500,6500,6500,6500,6500,6500,6500,26000
