+++ tests/fileTypeTests.py
@@ -114,7 +114,7 @@ class FileTypeTests(BaseTestValidator):
         """Test mixed job with some rows failing."""
         jobId = self.jobIdDict["programMixed"]
         self.passed = self.run_test(
-            jobId, 200, "finished", 20177, 4, "complete", 120, True)
+            jobId, 200, "finished", 21187, 4, "complete", 123, True)
 
 
     def test_award_fin_valid(self):
@@ -127,7 +127,7 @@ class FileTypeTests(BaseTestValidator):
         """Test mixed job with some rows failing."""
         jobId = self.jobIdDict["awardFinMixed"]
         self.passed = self.run_test(
-            jobId, 200, "finished", 16011, 6, "complete", 84, True)
+            jobId, 200, "finished", 17022, 6, "complete", 87, True)
 
     def test_award_valid(self):
         """Test valid job."""
