+++ tests/programActivityMixed.csv
@@ -1,11 +1,11 @@
 allocationtransferagencyidentifier,agencyidentifier,beginningperiodofavailability,endingperiodofavailability,availabilitytypecode,mainaccountcode,subaccountcode,programactivityname,programactivitycode,objectclass,bydirectreimbursablefundingsource,obligationsincurredbyprogramobjectclass_cpe,obligationsundeliveredordersunpaidtotal_fyb,obligationsundeliveredordersunpaidtotal_cpe,ussgl480100_undeliveredordersobligationsunpaid_fyb,ussgl480100_undeliveredordersobligationsunpaid_cpe,ussgl483100_undeliveredordersobligationstransferredunpaid_cpe,ussgl488100_upwardadjustmentsofprioryearundeliveredordersobligationsunpaid_cpe,obligationsdeliveredordersunpaidtotal_fyb,obligationsdeliveredordersunpaidtotal_cpe,ussgl490100_deliveredordersobligationsunpaid_fyb,ussgl490100_deliveredordersobligationsunpaid_cpe,ussgl493100_deliveredordersobligationstransferredunpaid_cpe,ussgl498100_upwardadjustmentsofprioryeardeliveredordersobligationsunpaid_cpe,grossoutlayamountbyprogramobjectclass_fyb,grossoutlayamountbyprogramobjectclass_cpe,grossoutlaysundeliveredordersprepaidtotal_fyb,grossoutlaysundeliveredordersprepaidtotal_cpe,ussgl480200_undeliveredordersobligationsprepaidadvanced_fyb,ussgl480200_undeliveredordersobligationsprepaidadvanced_cpe,ussgl483200_undeliveredordersobligationstransferredprepaidadvanced_cpe,ussgl488200_upwardadjustmentsofprioryearundeliveredordersobligationsprepaidadvanced_cpe,grossoutlaysdeliveredorderspaidtotal_fyb,grossoutlaysdeliveredorderspaidtotal_cpe,ussgl490200_deliveredordersobligationspaid_cpe,ussgl490800_authorityoutlayednotyetdisbursed_fyb,ussgl490800_authorityoutlayednotyetdisbursed_cpe,ussgl498200_upwardadjustmentsofprioryeardeliveredordersobligationspaid_cpe,ussgl487100_downwardadjustmentsofprioryearunpaidundeliveredordersobligationsrecoveries_cpe,ussgl497100_downwardadjustmentsofprioryearunpaiddeliveredordersobligationsrecoveries_cpe,ussgl487200_downwardadjustmentsofprioryearprepaidadvancedundeliveredordersobligationsrefundscollected_cpe,ussgl497200_downwardadjustmentsofprioryearpaiddeliveredordersobligationsrefundscollected_cpe,deobligationsrecoveriesrefundsdofprioryearbyprogramobjectclass_cpe
-19,72,0,0,X,306,0,School Grants,2,1255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,,113,0,Overseas Programs,2,1255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
 ,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
-19,72,0,0,X,306,0,School Grants,X,X,R,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X
-1900,7200,10000,10000,XFA,10950,9999,School Grants,20000,25.401,RR,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
-19,72,0,0,F,306,1,School Grants,2,3255,Q,5001,5002,5003,5004,5005,5006,5007,5008,5009,5010,5011,5012,5013,5014,5015,5016,5017,5018,5019,5020,5021,5022,5023,5024,5025,5026,5027,5028,5029,5030,5031,5032
-19,72,0,0,A,306,0,School Grants,2,1255,,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
-19,72,0,0,B,306,0,School Grants,2,1255,R,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
-19,72,0,0,X,306,0,School Grants,2,1255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
-19,72,0,0,X,306,0,School Grants,2,255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
-19,72,0,0,X,306,0,School Grants,2,255,R,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,,113,0,Overseas Programs,X,X,R,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X,X
+1900,7200,10000,10000,XFA,10950,9999,Overseas Programs,20000,25.401,RR,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2014,2016,F,999,1,Overseas Programs,2,3255,Q,5001,5002,5003,5004,5005,5006,5007,5008,5009,5010,5011,5012,5013,5014,5015,5016,5017,5018,5019,5020,5021,5022,5023,5024,5025,5026,5027,5028,5029,5030,5031,5032
+0,19,2016,2016,A,113,0,Overseas Programs,2,1255,,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,B,113,0,Overseas Programs,2,1255,R,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,,113,0,Overseas Programs,2,1255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,,113,0,Overseas Programs,2,255,D,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
+0,19,2016,2016,,113,0,Overseas Programs,2,255,R,4500,4500,4500,4500,1500,1000,1000,4500,4500,4500,1500,1000,1000,4500,4500,1000,1000,1000,400,300,100,3500,3500,1000,3500,1000,500,1000,1000,200,1000,3200
