+++ tests/runTests.py
@@ -28,16 +28,16 @@ def runTests(argv=''):
     # Create test suite
     suite = unittest.TestSuite()
 
-    #suite.addTests(unittest.makeSuite(LoginTests))
-    #suite.addTests(unittest.makeSuite(FileTests))
-    #suite.addTests(unittest.makeSuite(UserTests))
-    #suite.addTests(unittest.makeSuite(ValidatorTests))
-    #suite.addTests(unittest.makeSuite(JobTests))
+    suite.addTests(unittest.makeSuite(LoginTests))
+    suite.addTests(unittest.makeSuite(FileTests))
+    suite.addTests(unittest.makeSuite(UserTests))
+    suite.addTests(unittest.makeSuite(ValidatorTests))
+    suite.addTests(unittest.makeSuite(JobTests))
     suite.addTests(unittest.makeSuite(FileTypeTests))
 
     # to run a single test:
     #suite.addTest(FileTests('test_bad_quarter_or_month'))
-    #suite.addTest(FileTypeTests('test_award_valid'))
+    #suite.addTest(FileTypeTests('test_program_valid'))
 
     print("{} tests in suite".format(suite.countTestCases()))
